Refactor: libs폴더 제거 -> services & utils로 분리
