+++ src/hooks/useDndSensors.ts
@@ -1,0 +1,11 @@
+import { PointerSensor, useSensor, useSensors } from "@dnd-kit/core";
+
+export function useDndSensors() {
+  return useSensors(
+    useSensor(PointerSensor, {
+      activationConstraint: {
+        distance: 1,
+      },
+    }),
+  );
+}
